requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.4
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
django-tinymce4-lite==1.8.0
jsmin==2.2.2
django-tinymce==3.3.0
django-crispy-forms==1.12.0
